# docs/API_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BehavioralWell \u2014 Web + Android Shared Platform",
    "description": "Privacy-conscious multimodal behavioral risk detection, digital phenotyping, and early intervention API engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/telemetry/ingest": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Single Telemetry",
        "operationId": "ingest_single_telemetry_api_v1_telemetry_ingest_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Single Telemetry",
        "operationId": "ingest_single_telemetry_api_v1_telemetry_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/batch": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Batch Telemetry",
        "description": "Offline-first batch ingestion endpoint with idempotency deduplication.\nAllows Android client to queue local telemetry and upload upon reconnecting.",
        "operationId": "ingest_batch_telemetry_api_v1_telemetry_batch_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/status": {
      "get": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Get Telemetry Status",
        "operationId": "get_telemetry_status_api_v1_telemetry_status_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/telemetry/ingest": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Single Telemetry",
        "operationId": "ingest_single_telemetry_api_telemetry_ingest_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Single Telemetry",
        "operationId": "ingest_single_telemetry_api_telemetry_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/batch": {
      "post": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Ingest Batch Telemetry",
        "description": "Offline-first batch ingestion endpoint with idempotency deduplication.\nAllows Android client to queue local telemetry and upload upon reconnecting.",
        "operationId": "ingest_batch_telemetry_api_telemetry_batch_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/telemetry/status": {
      "get": {
        "tags": [
          "Telemetry & Processing"
        ],
        "summary": "Get Telemetry Status",
        "operationId": "get_telemetry_status_api_telemetry_status_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "usr_demo12345",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchTelemetryInput": {
        "properties": {
          "batch": {
            "items": {
              "$ref": "#/components/schemas/TelemetryInput"
            },
            "type": "array",
            "title": "Batch"
          }
        },
        "type": "object",
        "required": [
          "batch"
        ],
        "title": "BatchTelemetryInput"
      },
...
      "ContributorFactor": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "z_score": {
            "type": "number",
            "title": "Z Score"
          },
          "human_explanation": {
            "type": "string",
            "title": "Human Explanation"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "modality",
          "direction",
          "z_score",
          "human_explanation"
        ],
        "title": "ContributorFactor"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModalityBreakdown": {
        "properties": {
          "keyboard": {
            "type": "number",
            "title": "Keyboard"
          },
          "usage": {
            "type": "number",
            "title": "Usage"
          },
          "motion": {
            "type": "number",
            "title": "Motion"
          },
          "work": {
            "type": "number",
            "title": "Work"
          },
          "mobility": {
            "type": "number",
            "title": "Mobility"
          }
        },
        "type": "object",
        "required": [
          "keyboard",
          "usage",
          "motion",
          "work",
          "mobility"
        ],
        "title": "ModalityBreakdown"
      },
      "RiskAssessmentResponse": {
        "properties": {
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "stage": {
            "type": "integer",
            "title": "Stage"
          },
          "stage_label": {
            "type": "string",
            "title": "Stage Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "persistence_days": {
            "type": "integer",
            "title": "Persistence Days"
          },
          "top_contributors": {
            "items": {
              "$ref": "#/components/schemas/ContributorFactor"
            },
            "type": "array",
            "title": "Top Contributors"
          },
          "modality_scores": {
            "$ref": "#/components/schemas/ModalityBreakdown"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "stage",
          "stage_label",
          "confidence",
          "trend",
          "persistence_days",
          "top_contributors",
          "modality_scores",
          "timestamp"
        ],
        "title": "RiskAssessmentResponse"
      },
...
      "TelemetryInput": {
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "typing_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typing Speed"
          },
          "key_press_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Press Duration"
          },
          "pause_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Duration"
          },
          "correction_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Rate"
          },
          "screen_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screen Time"
          },
          "unlock_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unlock Count"
          },
          "night_usage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Night Usage"
          },
          "app_switch_frequency": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Switch Frequency"
          },
          "movement_intensity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movement Intensity"
          },
          "acceleration_variance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceleration Variance"
          },
          "stationary_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stationary Duration"
          },
          "task_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Accuracy"
          },
          "task_completion_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Completion Time"
          },
          "task_error_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Error Rate"
          },
          "speed_variance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Variance"
          },
          "route_variability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Variability"
          }
        },
        "type": "object",
        "title": "TelemetryInput"
      },
      "TelemetryStatusResponse": {
        "properties": {
          "total_records": {
            "type": "integer",
            "title": "Total Records"
          },
          "last_telemetry_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Telemetry At"
          },
          "modalities_collected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities Collected"
          }
        },
        "type": "object",
        "required": [
          "total_records",
          "last_telemetry_at",
          "modalities_collected"
        ],
        "title": "TelemetryStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}